completed frontend -> backend integration

+++ static/index.js
@@ -17,6 +17,8 @@ window.onload = function() {
         .addEventListener('click', deleteElementRequest);
     document.querySelector("#upload-button")
         .addEventListener('click', uploadFileRequest);
+    document.querySelector("#view-button")
+        .addEventListener('click', viewMoleculeRequest);
 };
 
 
@@ -72,6 +74,62 @@ function buildElement() {
 }
 
 
+// add element to element-list on client
+function appendElement(element) {
+    const newElement = document.createElement('option');
+    newElement.text = element;
+    newElement.id = `element-${element}`;
+
+    document.querySelector("#element-list").appendChild(newElement);
+}
+
+
+// add molecule to molecule-list on client
+function appendMolecule(molecule) {
+    const newMolecule = document.createElement('option');
+
+    // TODO: fix this text -  
+    newMolecule.text = molecule.name + " : " + molecule.atoms + " : " + molecule.bonds;
+    newMolecule.value = molecule.name;
+    newMolecule.id = `molecule-${molecule.name}`;
+
+    document.querySelector("#molecule-list").appendChild(newMolecule);
+}
+
+
+// remove element from element-list
+function removeElement(element) {
+    const elementList = document.querySelector("#element-list");
+    const optionToRemove = document.querySelector(`#element-${element}`);
+
+    elementList.removeChild(optionToRemove);
+}
+
+
+// build form data object with .sdf file data
+// return null if name field is empty or no file uploaded
+function buildFormData() {
+    const moleculeName = document.querySelector("#upload-molecule-name").value;
+    if (moleculeName === "") {
+        alert('Missing input in element name field');
+        return null;
+    }
+
+    const formData = new FormData();
+    const file = document.querySelector("#file").files[0];
+
+    if (!file) {
+        alert('No File Selected For Upload');
+        return null;
+    }
+
+    formData.append('name', moleculeName);
+    formData.append('file', file);
+
+    return formData;
+}
+
+
 // add element to database and append to list on client
 function addElementRequest() {
     const element = buildElement();
@@ -96,28 +154,6 @@ function addElementRequest() {
 }
 
 
-// add element to element-list on client
-function appendElement(element) {
-    const newElement = document.createElement('option');
-    newElement.text = element;
-    newElement.id = `element-${element}`;
-
-    document.querySelector("#element-list").appendChild(newElement);
-}
-
-
-// add molecule to molecule-list on client
-function appendMolecule(molecule) {
-    const newMolecule = document.createElement('option');
-
-    // TODO: fix this text
-    newMolecule.text = molecule.name + " : " + molecule.atoms + " : " + molecule.bonds;
-    newMolecule.id = `molecule-${molecule.name}`;
-
-    document.querySelector("#molecule-list").appendChild(newMolecule);
-}
-
-
 // load elements from db to element-list
 function loadElementsRequest() {
     fetch(BASE_URL + "/element_list", {method: 'GET'})
@@ -134,20 +170,12 @@ function loadMoleculesRequest() {
 }
 
 
-// remove element from element-list
-function removeElement(element) {
-    const elementList = document.querySelector("#element-list");
-    const optionToRemove = document.querySelector(`#element-${element}`);
-
-    elementList.removeChild(optionToRemove);
-}
-
-
 // delete element from database and from client
 function deleteElementRequest() {
     const elementName = document.querySelector("#element-list").value;
-
-    console.log(elementName);
+    if (elementName === "") {
+        return;
+    }
 
     fetch(BASE_URL + "/elements/" + elementName, {method: 'DELETE'})
     .then(res => {
@@ -157,28 +185,6 @@ function deleteElementRequest() {
             alert(UNEXPECTED_ERROR);
         }
     });
-}
-
-
-// build form data object with .sdf file data
-// return null if name field is empty or no file uploaded
-function buildFormData() {
-    const moleculeName = document.querySelector("#upload-molecule-name").value;
-    console.log(moleculeName);
-    if (moleculeName === "") {
-        alert('Missing input in element name field');
-        return null;
-    }
-
-    const formData = new FormData();
-    const file = document.querySelector("#file").files[0];
-
-    if (!file) {
-        alert('No File Selected For Upload');
-        return null;
-    }
-
-    formData.append('file', file);
 }
 
 
@@ -193,6 +199,21 @@ function uploadFileRequest() {
     .then(res => res.json())
     .then(res => {
         alert(res.message);
+
+        if (res.molecule) {
+            appendMolecule(res.molecule);
+        }
     })
     .catch(error => console.log(error));
+}
+
+
+// go to page to view selected molecule
+function viewMoleculeRequest() {
+    const molecule = document.querySelector("#molecule-list").value;
+    if (molecule === "") {
+        return;
+    }
+
+    window.location.href = `molecule/${molecule}`;
 }
